requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.18.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
tensorflow>=2.19.0
tf-keras>=2.19.0  # Required for DeepFace

# Face Detection
ultralytics>=8.0.0  # YOLOv8 for face detection
deepface>=0.0.79  # DeepFace for face verification

# Data handling
pyarrow>=12.0.0  # for parquet support
fastparquet>=0.8.0
huggingface-hub>=0.16.0

# Utilities
python-dotenv>=0.19.0
requests>=2.31.0
aiohttp>=3.8.0
tensorboard>=2.7.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Added from the code block
matplotlib>=3.7.0

# Added from the code block
streamlit>=1.24.0 